Tests/LibWeb/Text/expected/mathml-action-selection-style.txt

first[0]: block math
first[1]: none
second[0]: none
second[1]: block math
